requirements.txt
aiofiles==24.1.0
annotated-types==0.7.0
anyio==4.9.0
blinker==1.9.0
BTrees==6.1
certifi==2025.6.15
cffi==1.17.1
Chameleon==4.6.0
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
distro==1.9.0
dotenv==0.9.9
elevenlabs==2.5.0
faiss-cpu==1.11.0
fastapi==0.115.14
ffmpeg==1.4
ffmpy==0.6.0
filelock==3.18.0
Flask==3.1.1
fsspec==2025.5.1
gradio==5.35.0
gradio_client==1.10.4
grok==6.1
grokcore.annotation==5.0
grokcore.catalog==5.0
grokcore.chameleon==5.0
grokcore.component==5.0
grokcore.content==5.0
grokcore.formlib==5.0
grokcore.layout==5.0
grokcore.message==5.0
grokcore.security==5.0
grokcore.site==5.0
grokcore.traverser==5.0
grokcore.view==5.0
grokcore.viewlet==5.0
groovy==0.1.2
groq==0.29.0
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.2
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
martian==2.1
mdurl==0.1.2
mpmath==1.3.0
multipart==1.2.1
networkx==3.5
numpy==2.3.1
orjson==3.10.18
packaging==25.0
pandas==2.3.0
persistent==6.1.1
pillow==11.3.0
PyAudio==0.2.14
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-gettext==5.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
ruff==0.12.1
safehttpx==0.1.6
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.16.0
semantic-version==2.10.0
sentence-transformers==5.0.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
starlette==0.46.2
sympy==1.14.0
threadpoolctl==3.6.0
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.7.1
tqdm==4.67.1
transaction==5.0
transformers==4.53.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
websockets==15.0.1
Werkzeug==3.1.3
z3c.flashmessage==4.0
z3c.pt==5.1
zc.catalog==4.0
zc.lockfile==3.0.post1
ZConfig==4.2
zdaemon==5.2
ZODB==6.0.1
zodbpickle==4.2
zope.annotation==5.1
zope.app.appsetup==5.0
zope.app.publication==5.0
zope.app.wsgi==5.3
zope.authentication==5.1
zope.browser==3.1
zope.browserpage==5.1
zope.browserresource==5.2
zope.cachedescriptors==5.1
zope.catalog==5.1
zope.component==6.0
zope.configuration==6.0
zope.container==6.1
zope.contentprovider==6.0
zope.contenttype==5.2
zope.datetime==5.1
zope.deferredimport==5.0
zope.deprecation==5.1
zope.dottedname==6.0
zope.error==5.1
zope.errorview==2.0
zope.event==5.1
zope.exceptions==5.2
zope.filerepresentation==6.1
zope.formlib==6.1
zope.generations==5.1.0
zope.hookable==7.0
zope.i18n==5.2
zope.i18nmessageid==7.0
zope.index==7.0
zope.interface==7.2
zope.intid==5.1
zope.keyreference==6.1
zope.lifecycleevent==5.1
zope.location==5.1
zope.login==3.1
zope.minmax==2.3
zope.pagetemplate==5.1
zope.password==6.0
zope.principalregistry==5.1
zope.processlifetime==3.1
zope.proxy==6.1
zope.ptresource==5.1
zope.publisher==7.3
zope.schema==7.0.1
zope.security==7.3
zope.securitypolicy==5.1
zope.session==5.2
zope.site==5.1
zope.size==5.1
zope.tal==5.1
zope.tales==6.1
zope.testing==5.1
zope.traversing==5.1
zope.viewlet==5.1
